pyproject.toml
[project]
name = "mini-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "volcengine>=1.0.207",
    "volcengine-python-sdk>=4.0.35",
]
